third_party/WebKit/LayoutTests/http/tests/devtools/console/console-timestamp.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/console-test.js"></script>
<script>
function test() {
  // It is essential that we calculate timezone for this particular moment of time
  // otherwise the time zone offset could be different because of DST.
  var baseDate = Date.parse('2014-05-13T16:53:20.123Z');
  var tzOffset = new Date(baseDate).getTimezoneOffset() * 60 * 1000;
  var baseTimestamp = 1400000000000 + tzOffset;

  function addMessageWithFixedTimestamp(messageText, timestamp, type) {
    var message = new ConsoleModel.ConsoleMessage(
        TestRunner.runtimeModel,
        ConsoleModel.ConsoleMessage.MessageSource.Other,  // source
        ConsoleModel.ConsoleMessage.MessageLevel.Info,    // level
        messageText, type,
        undefined,                                      // url
        undefined,                                      // line
        undefined,                                      // column
        undefined,                                      // requestId
        undefined,                                      // parameters
        undefined,                                      // stackTrace
        timestamp || baseTimestamp + 123);              // timestamp: 2014-05-13T16:53:20.123Z
    TestRunner.consoleModel.addMessage(message, true);  // allowGrouping
  }

  TestRunner.addResult('Console messages with timestamps disabled:');
  addMessageWithFixedTimestamp('<Before> First Command', baseTimestamp + 789, ConsoleModel.ConsoleMessage.MessageType.Command);
  addMessageWithFixedTimestamp('<Before> First Result', baseTimestamp + 789, ConsoleModel.ConsoleMessage.MessageType.Result);
  addMessageWithFixedTimestamp('<Before>');
  addMessageWithFixedTimestamp('<Before>', baseTimestamp + 456);
  addMessageWithFixedTimestamp('<Before>');
  addMessageWithFixedTimestamp('<Before> Command', baseTimestamp, ConsoleModel.ConsoleMessage.MessageType.Command);
  addMessageWithFixedTimestamp('<Before> Result', baseTimestamp + 1, ConsoleModel.ConsoleMessage.MessageType.Result);

  ConsoleTestRunner.dumpConsoleMessages();

  TestRunner.addResult('Console messages with timestamps enabled:');
  Common.settingForTest('consoleTimestampsEnabled').set(true);

  addMessageWithFixedTimestamp('<After>', baseTimestamp + 1000);
  addMessageWithFixedTimestamp('<After>', baseTimestamp + 1000);
  addMessageWithFixedTimestamp('<After>', baseTimestamp + 1456);

  Common.settingForTest('consoleTimestampsEnabled').set(false);
  Common.settingForTest('consoleTimestampsEnabled').set(true);

  ConsoleTestRunner.dumpConsoleMessages();
  TestRunner.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>
Tests the console timestamp setting.
</p>
</body>
</html>
