third_party/WebKit/LayoutTests/http/tests/devtools/console/console-timestamp-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests the console timestamp setting.

Console messages with timestamps disabled:
<Before> First Command
"<Before> First Result"
3<Before>
<Before> Command
"<Before> Result"
Console messages with timestamps enabled:
16:53:20.789 <Before> First Command
16:53:20.789 "<Before> First Result"
16:53:20.123 <Before>
16:53:20.123 <Before>
16:53:20.456 <Before>
16:53:20.000 <Before> Command
16:53:20.001 "<Before> Result"
16:53:21.000 <After>
16:53:21.000 <After>
16:53:21.456 <After>
